fix ?q= redirect and refactor middleware
