+++ apps/sps-ordbokene-web/components/custom-pages/word-view.vue
@@ -110,7 +110,7 @@ useHead({
 
 definePageMeta({
     middleware: [
-      "pattern-redirect"
+      "simple-search"
     ]
   })
